# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface JellyfinAuthInput {
  /** Full URL of the Jellyfin server (e.g. https://jellyfin.example.com) */
  serverUrl: string;
  username: string;
  password: string;
}

export interface JellyfinCredentials {
  userId: string;
  accessToken: string;
  serverUrl: string;
  username: string;
}

export interface JellyfinLibrary {
  id: string;
  name: string;
  /** Jellyfin collection type: movies, tvshows, music, etc. */
  collectionType: string;
}

export interface JellyfinLibraryList {
  libraries: JellyfinLibrary[];
}

export interface ApiError {
  error: string;
  message: string;
}

export type GetJellyfinLibrariesParams = {
serverUrl: string;
userId: string;
accessToken: string;
};