# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: jellyfin
    description: Jellyfin server operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /jellyfin/auth:
    post:
      operationId: jellyfinAuth
      tags: [jellyfin]
      summary: Authenticate with a Jellyfin server
      description: Authenticates with the provided Jellyfin server and returns user credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JellyfinAuthInput"
      responses:
        "200":
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JellyfinCredentials"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Could not reach Jellyfin server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /jellyfin/libraries:
    get:
      operationId: getJellyfinLibraries
      tags: [jellyfin]
      summary: Get Jellyfin libraries
      description: Returns the list of media libraries from the Jellyfin server
      parameters:
        - name: serverUrl
          in: query
          required: true
          schema:
            type: string
        - name: userId
          in: query
          required: true
          schema:
            type: string
        - name: accessToken
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Libraries retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JellyfinLibraryList"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Could not reach Jellyfin server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    JellyfinAuthInput:
      type: object
      properties:
        serverUrl:
          type: string
          description: Full URL of the Jellyfin server (e.g. https://jellyfin.example.com)
        username:
          type: string
        password:
          type: string
      required:
        - serverUrl
        - username
        - password

    JellyfinCredentials:
      type: object
      properties:
        userId:
          type: string
        accessToken:
          type: string
        serverUrl:
          type: string
        username:
          type: string
      required:
        - userId
        - accessToken
        - serverUrl
        - username

    JellyfinLibrary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        collectionType:
          type: string
          description: "Jellyfin collection type: movies, tvshows, music, etc."
      required:
        - id
        - name
        - collectionType

    JellyfinLibraryList:
      type: object
      properties:
        libraries:
          type: array
          items:
            $ref: "#/components/schemas/JellyfinLibrary"
      required:
        - libraries

    ApiError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message