# lib/api-zod/src/generated/types/jellyfinAuthInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JellyfinAuthInput {
  /** Full URL of the Jellyfin server (e.g. https://jellyfin.example.com) */
  serverUrl: string;
  username: string;
  password: string;
}