# lib/api-zod/src/generated/types/jellyfinCredentials.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JellyfinCredentials {
  userId: string;
  accessToken: string;
  serverUrl: string;
  username: string;
}