# lib/api-zod/src/generated/types/jellyfinLibrary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JellyfinLibrary {
  id: string;
  name: string;
  /** Jellyfin collection type: movies, tvshows, music, etc. */
  collectionType: string;
}